requirements.txt
google-genai>=0.4.0
pillow>=10.0.0
requests>=2.31.0
google-auth>=2.23.0
google-api-python-client>=2.100.0
beautifulsoup4>=4.12.0
python-slugify>=8.0.0
lxml>=4.9.0